refactor: share code between node implementations
